Fix proxies in RemoteRepository.
We apparently have to configure the proxy by hand when we create our
RemoteRepository instances. The ProxySelector is only used when a repository
references another repository and Maven magicks up a new RemoteRepository
behind the scenes. Great design guys!

Another fun bit of Maven cleverness is that we cannot use the ProxySelector to
select a proxy when creating a RemoteRepository, because you have to already
*have* the RemoteRepository when you select a proxy. But you configure the
Proxy on a RemoteRepository with a builder and it is immutable. So we have to
create a throwaway RemoteRepository, select the right proxy, and then create a
new RemoteRepository with that proxy. Genius!